backend/requirements.txt
flask>=2.0
python-chess>=2.0
torch>=2.0
